Refactor: create file jupyter_util.py
